% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAchilles.R
\name{buildAchilles}
\alias{buildAchilles}
\title{Create achilles tables}
\usage{
buildAchilles(cdm, achillesId = NULL)
}
\arguments{
\item{cdm}{A cdm_reference object}

\item{achillesId}{A vector of achilles ids. You can also use "minimal",
"default" or "all" to point to a subset of analyses. Or also "person",
"observation period", ... to run the analyses refering to a certain category.}
}
\value{
The cdm_reference object with the achilles tables populated.
}
\description{
Create achilles tables
}
