% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAchillesTables.R
\name{buildAchillesTables}
\alias{buildAchillesTables}
\title{Build achilles tables using the data from the cdm object.}
\usage{
buildAchillesTables(cdm, achillesId = NULL)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{achillesId}{A vector of achilles ids. If NULL default analysis will be
used. Check the different analysis in \code{achillesAnalisisDetails}.}
}
\value{
The cdm_reference object with the achilles tables populated.
}
\description{
Build achilles tables using the data from the cdm object.
}
\examples{
\donttest{
dbName <- "GiBleed"
CDMConnector::requireEunomia(dbName)
con <- duckdb::dbConnect(duckdb::duckdb(), CDMConnector::eunomiaDir(dbName))
cdm <- CDMConnector::cdmFromCon(
  con = con, cdmSchema = "main", writeSchema = "main"
)

cdm <- buildAchillesTables(cdm = cdm)

}
}
