% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildPregnancyEpisodes.R
\name{buildPregnancyEpisodes}
\alias{buildPregnancyEpisodes}
\title{Build the \code{pregnancy_episodes} extenssion table}
\usage{
buildPregnancyEpisodes(cdm)
}
\arguments{
\item{cdm}{A cdm_reference object.}
}
\value{
A cdm_reference object with the \code{pregnancy_episodes} table.
}
\description{
Build the \code{pregnancy_episodes} extenssion table
}
