% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseRecords.R
\name{collapseRecords}
\alias{collapseRecords}
\title{Collapse records of a \code{cdm_table} into episodes.}
\usage{
collapseRecords(x, startDate, endDate, by, gap = 0L, name = NULL)
}
\arguments{
\item{x}{A \code{cdm_table} object.}

\item{startDate}{Column in \code{x} that points to the start date of the record.}

\item{endDate}{Column in \code{x} that point to the end date of the record.}

\item{by}{Columns in \code{x} that aggregate the records.}

\item{gap}{Integer; distance allowed between two consecutive records to be
collapsed.}

\item{name}{Name of the new \code{cdm_table} created.}
}
\value{
The \code{x} \code{cdm_table} with the records collapsed.
}
\description{
Collapse records of a \code{cdm_table} into episodes.
}
