% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OmopConstructor-package.R
\docType{package}
\name{OmopConstructor-package}
\alias{OmopConstructor}
\alias{OmopConstructor-package}
\title{OmopConstructor: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-pharmacoepi.github.io/OmopConstructor/}
}

}
\author{
\strong{Maintainer}: Martí Català \email{marti.catalasabate@ndorms.ox.ac.uk} (\href{https://orcid.org/0000-0003-3308-9905}{ORCID})

}
\keyword{internal}
