% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateObservationPeriod.R
\name{generateObservationPeriod}
\alias{generateObservationPeriod}
\title{Generate \code{observation_period} table from data recorded in the
\code{cdm_reference}.}
\usage{
generateObservationPeriod(
  cdm,
  collapseEra = Inf,
  persistenceWindow = Inf,
  censorDate = Sys.time(),
  censorAge = 120L,
  recordsFrom = c("drug_exposure", "condition_occurrence", "procedure_occurrence",
    "visit_occurrence", "device_exposure", "measurement", "observation", "death"),
  periodTypeConceptId = 32817L
)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}

\item{collapseEra}{Distance between records to be collapsed.}

\item{persistenceWindow}{Number of days added at the end of an observation
period as persistence window.}

\item{censorDate}{Date to censor all followup for individuals.}

\item{censorAge}{Age to censor individuals if they reach a certain age.}

\item{recordsFrom}{Tables to retrieve observation records from.}

\item{periodTypeConceptId}{Choose the observation_period_type_concept_id that
best represents how the period was determined.
\href{https://athena.ohdsi.org/search-terms/terms?domain=Type+Concept&standardConcept=Standard&page=1&pageSize=15&query=}{Accepted Concepts}.}
}
\value{
The \code{cdm_reference} object with a new \code{observation_period}.
}
\description{
Generate \code{observation_period} table from data recorded in the
\code{cdm_reference}.
}
